+++ src/components/main-sidebar.tsx
@@ -8,6 +8,8 @@ import {
   SidebarMenuBadge,
   SidebarMenuButton,
   SidebarMenuItem,
+  SidebarRail,
+  useSidebar,
 } from "@/components/ui/sidebar"
 import { SIDEBAR_FOOTER_LINKS, SIDEBAR_LINKS } from "@/lib/constants/links";
 import { Link, useLocation } from "react-router"
@@ -17,6 +19,7 @@ import { useSettings } from "@/context/settings";
 import { useTranslation } from "react-i18next";
 import { SidebarLink } from "@/lib/types/enums";
 import { useQuarantineCount } from "@/context/quarantine-count";
+import { useIsMobile } from "@/hooks/use-mobile";
 
 export default function MainSidebar(){
      const location = useLocation();
@@ -24,10 +27,12 @@ export default function MainSidebar(){
      const {quarantineCount} = useQuarantineCount()
      const sidebarLinks = useMemo(()=>!settings.enableSchedulerUI ? SIDEBAR_LINKS.filter(val=>val.href!=="/scheduler") : SIDEBAR_LINKS,[settings.enableSchedulerUI])
      const {t} = useTranslation();
+     const {state} = useSidebar()
+     const isMobile = useIsMobile()
      return (
-          <Sidebar>
+          <Sidebar collapsible="icon">
                <SidebarHeader>
-                    <Logo width={1500} height={400}/>
+                    <Logo width={isMobile ? 370 : 240} height={isMobile ? 100 : 64} state={state}/>
                </SidebarHeader>
                <SidebarContent>
                     <SidebarGroup>
@@ -46,16 +51,15 @@ export default function MainSidebar(){
                     </SidebarGroup>
                </SidebarContent>
                <SidebarFooter>
-                    <SidebarGroup>
-                         <SidebarMenu>
-                              {SIDEBAR_FOOTER_LINKS.map(({Icon,name,href},i)=>(
-                                   <SidebarMenuButton isActive={location.pathname===href} asChild key={`${name}-${i}`}>
-                                        <Link to={href} className="text-muted-foreground"><Icon className="text-primary dark:text-accent-foreground"/> {t(`sidebar.${name}`)}</Link>
-                                   </SidebarMenuButton>
-                              ))}
-                         </SidebarMenu>
-                    </SidebarGroup>
+                    <SidebarMenu>
+                         {SIDEBAR_FOOTER_LINKS.map(({Icon,name,href},i)=>(
+                              <SidebarMenuButton isActive={location.pathname===href} asChild key={`${name}-${i}`}>
+                                   <Link to={href} className="text-muted-foreground"><Icon className="text-primary dark:text-accent-foreground"/> {t(`sidebar.${name}`)}</Link>
+                              </SidebarMenuButton>
+                         ))}
+                    </SidebarMenu>
                </SidebarFooter>
+               <SidebarRail/>
           </Sidebar>
      )
 }